refactor(InputView): InputView가 정말 Input만을 관리하도록 변경
책임을 각 당첨번호, 보너스번호 객체에 넘기니 Input 관련 코드가 깔끔해짐
